+++ Changes.md
@@ -88,6 +88,8 @@ Unreleased
   ([#1275](https://github.com/melange-re/melange/pull/1275))
 - core: improve generated code formatting for `throw` and `return` statements,
   JS objects ([#1286](https://github.com/melange-re/melange/pull/1286))
+- core: improve generated code formatting for empty return and continue
+  statements ([#1288](https://github.com/melange-re/melange/pull/1288))
 
 4.0.1 2024-06-07
 ---------------